requirements.txt
Django==4.0.3
gunicorn==20.1.0
boto3==1.21.16
django-dotenv==1.4.2
django-storages==1.12.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
djangorestframework_simplejwt==5.1.0
